+++ apps/zui/src/app/features/sidebar/pools-section/empty.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import {Icon} from "src/components/icon"
-import EmptySection from "src/js/components/common/EmptySection"
+import EmptySection from "src/components/empty-section"
 import styled from "styled-components"
 
 export const Empty = styled(EmptySection).attrs({
