+++ apps/zui/src/components/link.tsx
@@ -3,7 +3,7 @@ import {invoke} from "src/core/invoke"
 
 type Props = {href?: string; children: any; onClick?: Function}
 
-export default function Link({href, onClick, children}: Props) {
+export function Link({href, onClick, children}: Props) {
   // Anchors can be passed through
   if (href?.startsWith("#")) return <a href={href}>{children}</a>
